feat: SM-14 enum 타입으로 상수화 시킨 값을 as const를 사용한 객체로 변경하여 적용
