+++ src/utils/constants/color.ts
@@ -1,7 +1,11 @@
-enum THEME_COLOR {
-  Primary = 'grey',
-  Dark = 'black',
-  Light = 'white'
-}
+const THEME_COLOR = {
+  primary: 'grey',
+  dark: 'black',
+  light: 'white'
+} as const;
+
+type ThemeColorKeys = keyof typeof THEME_COLOR;
+type IThemeColor = typeof THEME_COLOR[ThemeColorKeys];
 
 export { THEME_COLOR };
+export type { IThemeColor };
