--- src/utils/constants/size.ts
@@ -1,16 +1,0 @@
-const MODAL_SIZE = {
-  lg: {
-    width: '1000px',
-    height: '1000px'
-  },
-  md: {
-    width: '500px',
-    height: '500px'
-  },
-  sm: {
-    width: '100px',
-    height: '100px'
-  }
-};
-
-export { MODAL_SIZE };
